add pngquant to tbbl

+++ tbbl/res.py
@@ -9,6 +9,7 @@
 import os
 import subprocess
 import shutil
+import tempfile
 from zrong import (slog, ZrongError)
 from zrong.base import (list_dir, copy_dir, get_files)
 import zrong.lua as lua
@@ -331,11 +332,61 @@ class ResBase(object):
     def oth(self, oth):
         self._commonProcess(oth, 'oth')
 
+    def ani(self, ani, gen_def):
+        default, special = self._getSpecialDir('ani')
+        if not default:
+            return
+
+        if special:
+            default = self._merge2Tmp(default, special, ani)
+        aniDef = AniDef(self.conf, default, ani, gen_def)
+        if not aniDef.checkAni():
+            return
+
+        outputDir = self.conf.getClientPath('res', 'ani')
+        self._rebuildDir(outputDir)
+
+        sp = aniDef.getAnisInSpriteSheet()
+        if sp:
+            print_sep('Start to copy existing animations.')
+            for adir in sp:
+                png = adir + '.png'
+                spng = os.path.join(default, png)
+                dpng = os.path.join(outputDir, png)
+                shutil.copyfile(spng, dpng)
+                self._print_copy(spng, dpng)
+
+                plist = adir + '.plist'
+                splist = os.path.join(default, plist)
+                dplist = os.path.join(outputDir, plist)
+                shutil.copyfile(splist, dplist)
+                self._print_copy(splist, dplist)
+            print_sep('Copy existing animations has done.', False)
+
+        defs = aniDef.getAniDefs()
+        if defs:
+            print_sep('Start to copy existing ani_def files.')
+            for anidef in defs:
+                sanidef = os.path.join(default, anidef)
+                danidef = os.path.join(outputDir, anidef)
+                shutil.copyfile(sanidef, danidef)
+                self._print_copy(sanidef, danidef)
+            print_sep('Copy existing ani_def files has done.', False)
+
+        print_sep('Start to convert pieces to animations and generate ani_def files.')
+        for adir in aniDef.getAnisInPiece():
+            self._convertSS(default, outputDir, adir)
+            aniDef.generateADef(adir)
+        print_sep('Converting and generating has done.', False)
+
     def test(self):
         res = self.conf.getGit('resource', 'path')
         sourcedir = os.path.join(res, 'test')
         targetdir = self.conf.getClientPath('res', 'test')
         self._print_copy(sourcedir, targetdir)
+        if not os.path.exists(sourcedir):
+            slog.error('No such file or directory: [%s].'%sourcedir)
+            return
         if os.path.exists(targetdir):
             shutil.rmtree(targetdir)
         shutil.copytree(sourcedir, targetdir)
@@ -383,7 +434,7 @@ class ResBase(object):
     def build(self):
         noAnyArgs = True
         if self.args.plst != None:
-            self.plist(self.args.plst)
+            self.plst(self.args.plst)
             noAnyArgs = False
         if self.args.pdir != None:
             self.pdir(self.args.pdir)
@@ -407,7 +458,7 @@ class ResBase(object):
             self.test()
             noAnyArgs = False
         if self.args.gettext:
-            self.gettext(sef.args.gettext)
+            self.gettext(self.args.gettext)
             noAnyArgs = False
 
         return noAnyArgs
